// web-ANGULAR/dfps-web-impact/src/app/modules/financial/contract/model/ContractServiceList.ts
export interface ContractServicesList {
    contractId: number;
    contractWorkerId: number;
    contractServiceId: number;
    period: number;
    version: number;
    lineItem: number;
    serviceCode: string;
    paymentType: string;
    isNewRow: string;
    unitType: string;
    federalMatch: number;
    localMatch: number;
    unitRate: number;
    adminAllUsedAmount: number;
    equipmentAmount: number;
    equipmentUsedAmount: number;
    fringeBenefitAmount: number;
    fringeBenefitUsedAmount: number;
    offsetItemUsedAmount: number;
    otherAmount: number;
    otherUsedAmount: number;
    salaryAmount: number;
    salaryUsedAmount: number;
    supplyAmount: number;
    supplyUsedAmount: number;
    travelAmount: number;
    travelUsedAmount: number;
    unitRateAmount: number;
    unitRateUsedAmount: number;
    totalUnits: number;
    usedUnits: number;
    lastUpdatedDate: string;
    pageMode: number;
    countyCodes: number;
    totalAmount: number;
    budgetBalanceAmount: number;
    scorContractNumber: string;
    paymentTypeCode: string;
}